pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["krivitskyanton <krivitsky.anton@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" }
]

[tool.poetry.scripts]
gendiff = "gendiff.gendiff:main"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
